# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: example123@gmail.com
        type: string
    required:
      - email
    type: object
...
  dto.JWT:
    properties:
      token:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: example123@gmail.com
        type: string
      password:
        example: example123
        minLength: 8
        type: string
    required:
      - email
      - password
    type: object
  dto.LoginResponse:
    properties:
      data:
        $ref: "#/definitions/dto.JWT"
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      confirm_password:
        example: example123
        type: string
      email:
        example: example123@gmail.com
        type: string
      fullname:
        example: John Doe
        minLength: 3
        type: string
      password:
        example: example123
        minLength: 8
        type: string
    required:
      - confirm_password
      - email
      - fullname
      - password
    type: object
  dto.RegisterResponse:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
  dto.ResponseError:
    properties:
      errors:
        example: failed get data
        type: string
      message:
        example: Failed get data
        type: string
      status:
        example: Error
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
  dto.UpdateForgotPasswordRequest:
    properties:
      confirm_password:
        example: example123
        type: string
      otp_code:
        example: "123456"
        maxLength: 6
        type: string
      password:
        example: example123
        minLength: 8
        type: string
    required:
      - confirm_password
      - otp_code
      - password
    type: object
...
host: 192.168.50.221:8080
info:
  contact: {}
  description: Solid Coffee Backend RESTful API
  title: Solid Coffee Backend
  version: "1.0"
paths:
...
  /auth/:
    post:
      consumes:
        - application/json
      description: Authenticate user with email and password
      parameters:
        - description: Login credentials
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.LoginResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: User login
      tags:
        - Auth
  /auth/forgot-password/:
    post:
      consumes:
        - application/json
      description: Send OTP to registered email
      parameters:
        - description: Forgot password request
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.ForgotPasswordRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Request OTP for forgot password
      tags:
        - Auth
  /auth/forgot-password/update:
    post:
      consumes:
        - application/json
      description: Verify OTP and update user password
      parameters:
        - description: Update forgot password
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.UpdateForgotPasswordRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Update password using OTP
      tags:
        - Auth
  /auth/new/:
    post:
      consumes:
        - application/json
      description: Create a new user account
      parameters:
        - description: User registration data
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.RegisterRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/dto.RegisterResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Register new user
      tags:
        - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"